+++ lib/launcher.js
@@ -86,16 +86,16 @@ class AppiumLauncher {
         return ret;
     }
 
-    _keyValueToCliArgs (keyValue) {
-        if (Array.isArray(keyValue)) {
+    _keyValueToCliArgs (args) {
+        if (Array.isArray(args)) {
             // Note:
             // If specified as array, this plugin assumes it as the string list of command line arguments.
-            return keyValue;
+            return args;
         }
 
         const ret = [];
-        for (let key in keyValue) {
-            const value = keyValue[key];
+        for (let key in args) {
+            const value = args[key];
             if (typeof value === 'boolean' && !value) {
                 continue;
             }
@@ -110,4 +110,4 @@ class AppiumLauncher {
     }
 }
 
-module.exports = new AppiumLauncher();
+export default AppiumLauncher;
