+++ package.json
@@ -8,7 +8,11 @@
   },
   "scripts": {
     "lint": "eslint .",
-    "test": "wdio test/wdio.conf.js"
+    "build": "grunt build",
+    "prepublish": "npm prune && npm run build",
+    "test": "npm run test:unit && npm run test:integration",
+    "test:unit": "mocha --compilers js:babel-register --recursive test/unit",
+    "test:integration": "wdio test/integration/wdio.conf.js"
   },
   "repository": {
     "type": "git",
