+++ test/integration/smoke_test.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const assert = require('assert');
+
+describe('appium-service', function () {
+    it('should run without an error', function () {
+        assert(browser);
+    });
+});
