--- test/smoke_test.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const assert = require('assert');
-
-describe('appium-service', function() {
-    it ('should run without an error', function() {
-        assert(browser);
-    });
-});
